+++ app/(admin)/admin/layout.tsx
@@ -1,5 +1,8 @@
 import { AdminSideBar } from '@/app/(admin)/admin/AdminSideBar';
 import { Header } from '@/components/common/Header';
+import { MobileNav } from '@/components/mobileUI/MobileNav';
+import { cn } from '@/lib/utils';
+import { getSessionUserServerAction } from '@/modules/user/domain/usecases/getSessionUserServerAction';
 import {
   CreditCard,
   Home,
@@ -65,21 +68,44 @@ const sidebarNavItems = [
   },
 ];
 
-export default function AdminLayout({
+export default async function AdminLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
+  const user = await getSessionUserServerAction();
   return (
-    <div className="pt-16">
-      <Header />
-      <div className="flex min-h-[calc(100vh-4rem)]">
-        <div className="fixed left-0 hidden h-full lg:block w-56 overflow-auto">
+    <div className="">
+      <Header user={user} />
+      <div className="flex h-auto min-h-[calc(100vh-4rem)]">
+        <div
+          className={cn(
+            'absolute left-0 z-40 w-56 h-auto pt-16 overflow-auto bg-background',
+            'hidden landscape:block', // 모바일 가로 모드일때 보이기
+            'lg:fixed lg:h-full' // 데스크탑부턴 헤더가 감춰지지 않음으로 fixed로 고정
+          )}
+        >
           <AdminSideBar items={sidebarNavItems} />
         </div>
-        <main className="w-full lg:pl-56">
-          <div className="h-full lg:border-l">{children}</div>
+        <main className="w-full lg:pl-56 landscape:pl-56">
+          <div
+            className={cn(
+              'h-full pb-40 pt-16', // 모바일은 BottomNav(h-16) + Player(h-16) = h-32
+              'landscape:border-l landscape:pb-28' // 가로 모드부터는 Player(h-20) = h-20
+            )}
+          >
+            {children}
+          </div>
         </main>
+
+        <div
+          className={cn(
+            'fixed bottom-0 w-full z-50 bg-foreground',
+            'landscape:hidden' // 모바일 가로 모드 사이즈부터는 BottomNav 사용 X
+          )}
+        >
+          <MobileNav />
+        </div>
       </div>
     </div>
   );
