+++ app/(main)/account/billing/page.tsx
@@ -24,18 +24,6 @@ const pricingBundle = [
       '전체 트랙 스트리밍',
       'JAZZiT+ 전용 콘텐츠',
       '코드 진행 악보 제공',
-    ],
-  },
-  {
-    badge: 'pro',
-    title: '프로',
-    description: '',
-    price: 16900,
-    isSelected: false,
-    benefits: [
-      '전체 트랙 스트리밍',
-      'JAZZiT+ 전용 콘텐츠',
-      '코드 진행 악보 제공',
       '모든 컨텐츠 다운로드',
       '개별 트랙 제공',
     ],
@@ -62,7 +50,7 @@ export default async function AccountBillingPage() {
         <h3 className="text-2xl font-medium">월 요금제</h3>
       </div>
       <Separator />
-      <div className="grid grid-cols-3 gap-4">
+      <div className="grid grid-cols-1 landscape:grid-cols-2 gap-4">
         {pricingBundle.map((pricing, idx) => {
           return (
             <PricingCard
