+++ app/(main)/layout.tsx
@@ -1,13 +1,6 @@
 import { Header } from '@/components/common/Header';
 import { MainSideBar } from '@/app/(main)/MainSideBar';
-import {
-  Heart,
-  LayoutGrid,
-  ListMusic,
-  Monitor,
-  Music,
-  Users,
-} from 'lucide-react';
+import { Heart, LayoutGrid, Monitor, Music, Users } from 'lucide-react';
 import { MobileNav } from '@/components/mobileUI/MobileNav';
 import { getSessionUserServerAction } from '@/modules/user/domain/usecases/getSessionUserServerAction';
 import { cn } from '@/lib/utils';
@@ -43,29 +36,6 @@ const userSideBarNavItems = [
   },
 ];
 
-const mobileNavItems = [
-  {
-    href: '/',
-    icon: <LayoutGrid className="h-4 w-4" />,
-    title: '탐색',
-  },
-  {
-    href: '/sounds',
-    icon: <Music className="h-4 w-4" />,
-    title: '트랙',
-  },
-  {
-    href: '/community',
-    icon: <Users className="h-4 w-4" />,
-    title: '커뮤니티',
-  },
-  {
-    href: '/playlists',
-    icon: <ListMusic className="h-4 w-4" />,
-    title: '플레이리스트',
-  },
-];
-
 export default async function MainLayout({
   children,
 }: {
@@ -74,12 +44,12 @@ export default async function MainLayout({
   const user = await getSessionUserServerAction();
 
   return (
-    <div className="pt-16">
+    <div className="">
       <Header user={user} />
       <div className="flex h-auto min-h-[calc(100vh-4rem)]">
         <div
           className={cn(
-            'absolute left-0 z-40 w-56 h-auto overflow-auto bg-background',
+            'absolute left-0 z-40 w-56 h-auto pt-16 overflow-auto bg-background',
             'hidden landscape:block', // 모바일 가로 모드일때 보이기
             'lg:fixed lg:h-full' // 데스크탑부턴 헤더가 감춰지지 않음으로 fixed로 고정
           )}
@@ -92,7 +62,7 @@ export default async function MainLayout({
         <main className="w-full lg:pl-56 landscape:pl-56">
           <div
             className={cn(
-              'h-full pb-40', // 모바일은 BottomNav(h-16) + Player(h-16) = h-32
+              'h-full pb-40 pt-16', // 모바일은 BottomNav(h-16) + Player(h-16) = h-32
               'landscape:border-l landscape:pb-28' // 가로 모드부터는 Player(h-20) = h-20
             )}
           >
@@ -106,7 +76,7 @@ export default async function MainLayout({
             'landscape:hidden' // 모바일 가로 모드 사이즈부터는 BottomNav 사용 X
           )}
         >
-          <MobileNav mobileNavItems={mobileNavItems} />
+          <MobileNav />
         </div>
       </div>
     </div>
